+++ libs/react/src/components/molecules/surveyQuestionContainer/surveyQuestionContainer.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { SurveyInput } from '../index';
 import { cloneDeep, findIndex, forEach, forOwn } from 'lodash';
 import { IButtonProps, SurveyActiveKeyType, SurveyAdditionalContext, SurveyPayloadType, SurveySectionType } from '@coldpbc/interfaces';
@@ -31,8 +31,6 @@ const _SurveyQuestionContainer = ({ activeKey, setActiveKey, submitSurvey, surve
     enterDone: 'transition ease-out duration-200 transform translate-y-0',
     exitActive: 'transition ease-in duration-200 transform translate-y-full',
   };
-  const [activeQuestion, setActiveQuestion] = React.useState<JSX.Element | undefined>(undefined);
-  const [additionalContextQuestion, setAdditionalContextQuestion] = React.useState<JSX.Element | undefined>(undefined);
   const [transitionClassNames, setTransitionClassNames] = React.useState<any>(nextQuestionTransitionClassNames);
   const { definition, id, name } = surveyData;
   const { sections } = definition;
@@ -592,9 +590,9 @@ const _SurveyQuestionContainer = ({ activeKey, setActiveKey, submitSurvey, surve
       }
     }
     if (condition) {
-      setAdditionalContextQuestion(getQuestionForKey(activeKey, true));
+      return getQuestionForKey(activeKey, true);
     } else {
-      setAdditionalContextQuestion(undefined);
+      return undefined;
     }
   };
 
@@ -659,27 +657,20 @@ const _SurveyQuestionContainer = ({ activeKey, setActiveKey, submitSurvey, surve
     })
     .flat();
 
-  const getActiveQuestion = () => {
-    const question = questions.find(question => {
-      return question.props.input_key === activeKey.value;
-    });
-    // show additional context
-    checkAdditionalContext(activeKey);
-    setActiveQuestion(question);
-  };
+  const question = questions.find(question => {
+    return question.props.input_key === activeKey.value;
+  });
 
-  useEffect(() => {
-    getActiveQuestion();
-  }, [activeKey, surveyData]);
+  const additionalContextQuestion = checkAdditionalContext(activeKey);
 
-  if (activeQuestion !== undefined) {
+  if (question !== undefined) {
     return (
       <div className={'w-full h-full relative flex items-center justify-center overflow-hidden pb-[93px]'}>
         <SwitchTransition>
-          <CSSTransition key={activeQuestion.props.input_key} timeout={150} classNames={transitionClassNames}>
+          <CSSTransition key={question.props.input_key} timeout={150} classNames={transitionClassNames}>
             <div className={'h-full w-full flex items-center justify-center px-[139px] shortScreen:px-[32px] shortWideScreen:px-[139px]'}>
               <div className={'w-full space-y-6'}>
-                {activeQuestion}
+                {question}
                 {additionalContextQuestion}
               </div>
             </div>